+++ src/Core/Components/DateTime/FluentDatePicker.razor
@@ -44,7 +44,8 @@
 @if (IsFluentUIStyle)
 {
     <FluentPopover AnchorId="@Id" @bind-Opened="@Opened">
-        <FluentCalendar TValue="TValue" @ref="_calendar"
+        <FluentCalendar @ref="_calendar"
+                        TValue="TValue"
                         Culture="@Culture"
                         View="@View"
                         DayFormat="@DayFormat"
@@ -54,6 +55,6 @@
                         Value="@Value"
                         ValueChanged="@OnSelectedDateAsync"
                         DaysTemplate="@DaysTemplate"
-                        PickerMonthChanged="@(month => PickerMonthChanged.InvokeAsync(month.ConvertToDateTime() ?? DateTimeProvider.Today))" />
+                        PickerMonthChanged="@PickerMonthChangedHandlerAsync" />
     </FluentPopover>
 }
