+++ src/Core/Localization/LanguageResource.resx
@@ -300,4 +300,7 @@
   <data name="DatePicker_ButtonTitle" xml:space="preserve">
     <value>Show date picker</value>
   </data>
+  <data name="Calendar_FieldMustBeADate" xml:space="preserve">
+    <value>The {0} field must be a date.</value>
+  </data>
 </root>
